reordena context
