+++ App.js
@@ -1,34 +1,33 @@
 import { NativeBaseProvider, extendTheme } from "native-base";
 import { StatusBar } from "expo-status-bar";
 import Routes from "./src/Routes";
-import { ContextProvider } from "./src/Providers/context";
+import { ContextProvider, SearchProvider } from "./src/Providers/context";
 
 export default function App() {
-    const theme = extendTheme({
-        colors: {
-            primary: {
-                50: "#f5f3ff",
-                100: "#ede9fe",
-                200: "#ddd6fe",
-                300: "#c4b5fd",
-                400: "#a78bfa",
-                500: "#8b5cf6",
-                600: "#7c3aed",
-                700: "#6d28d9",
-                800: "#5b21b6",
-                900: "#4c1d95",
-            },
-        },
-    });
-    return (
-        <ContextProvider>
-            <NativeBaseProvider theme={theme}>
-                <StatusBar
-                    barStyle="dark-content"
-                    backgroundColor="transparent"
-                />
-                <Routes />
-            </NativeBaseProvider>
-        </ContextProvider>
-    );
+  const theme = extendTheme({
+    colors: {
+      primary: {
+        50: "#f5f3ff",
+        100: "#ede9fe",
+        200: "#ddd6fe",
+        300: "#c4b5fd",
+        400: "#a78bfa",
+        500: "#8b5cf6",
+        600: "#7c3aed",
+        700: "#6d28d9",
+        800: "#5b21b6",
+        900: "#4c1d95",
+      },
+    },
+  });
+  return (
+    <ContextProvider>
+      <SearchProvider>
+        <NativeBaseProvider theme={theme}>
+          <StatusBar barStyle="dark-content" backgroundColor="transparent" />
+          <Routes />
+        </NativeBaseProvider>
+      </SearchProvider>
+    </ContextProvider>
+  );
 }
