+++ src/screens/private/Search/components/Inputs/index.jsx
@@ -13,12 +13,12 @@ import {
 } from "native-base";
 
 import Icon from "react-native-vector-icons/FontAwesome5";
-import { SearchContext } from "../../context";
 import { useNavigation } from "@react-navigation/native";
+import { SearchContext } from "../../../../../Providers/context";
 
 export default function Inputs({ props }) {
   const { toggleSearch, isSearching, fullParty, resetSearch } =
-  useContext(SearchContext);
+    useContext(SearchContext);
   const navigation = useNavigation();
 
   const handleConfirm = async () => {
@@ -27,39 +27,43 @@ export default function Inputs({ props }) {
 
   return (
     <Column
-      w={"95%"}
+      w={"100%"}
       borderWidth={1}
       borderColor={"primary.700"}
-      borderBottomRadius={8}
       backgroundColor={"white"}
       pt={7}
       px={2}
       pb={2}
     >
       <StatusBar />
-      <Text color={"primary.700"}>Origem</Text>
-      <Input
-        leftElement={
-          <NBIcon
-            as={<Icon name={"map-pin"} />}
-            size={15}
-            ml={2}
-            color={"primary.700"}
+      {isSearching && !fullParty ? null : (
+        <>
+          <Text color={"primary.700"}>Origem</Text>
+          <Input
+            h={12}
+            leftElement={
+              <NBIcon
+                as={<Icon name={"map-pin"} />}
+                size={15}
+                ml={2}
+                color={"primary.700"}
+              />
+            }
           />
-        }
-      />
-      <Divider my={2} />
-      <Text color={"primary.700"}>Destino</Text>
-      <Input
-        leftElement={
-          <NBIcon
-            as={<Icon name={"map-pin"} />}
-            size={15}
-            ml={2}
-            color={"primary.700"}
+          <Divider my={2} />
+          <Text color={"primary.700"}>Destino</Text>
+          <Input
+            leftElement={
+              <NBIcon
+                as={<Icon name={"map-pin"} />}
+                size={15}
+                ml={2}
+                color={"primary.700"}
+              />
+            }
           />
-        }
-      />
+        </>
+      )}
 
       {fullParty ? (
         <Column w={"100%"} mt={2}>
@@ -67,11 +71,21 @@ export default function Inputs({ props }) {
             <Button
               w={"40%"}
               colorScheme={"error"}
-              onPress={() => resetSearch()}
+              onPress={() => {
+                console.log("cancelar");
+                resetSearch();
+              }}
             >
               Cancelar
             </Button>
-            <Button w={"40%"} colorScheme={"success"} onPress={handleConfirm}>
+            <Button
+              w={"40%"}
+              colorScheme={"success"}
+              onPress={() => {
+                console.log("confirmar");
+                handleConfirm();
+              }}
+            >
               Confirmar
             </Button>
           </Button.Group>
@@ -82,6 +96,7 @@ export default function Inputs({ props }) {
           backgroundColor={"primary.700"}
           w={"100%"}
           onPress={() => {
+            console.log('pressed')
             toggleSearch();
           }}
         >
